Extract GetConsensusHash() into consensushash.h/cpp
