+++ src/omnicore/rules.h
@@ -182,8 +182,6 @@ bool IsAllowedOutputType(int whichType, int nBlock);
 /** Checks, if the transaction type and version is supported and enabled. */
 bool IsTransactionTypeAllowed(int txBlock, uint32_t txProperty, uint16_t txType, uint16_t version);
 
-/** Obtains a hash of all balances to use for consensus verification and checkpointing. */
-uint256 GetConsensusHash();
 /** Compares a supplied block, block hash and consensus hash against a hardcoded list of checkpoints. */
 bool VerifyCheckpoint(int block, const uint256& blockHash);
 }
